+++ pkg/namespace/delete.go
