+++ pkg/namespace/delete_test.go
