+++ pkg/namespace/error.go
