+++ pkg/namespace/resource.go
